WiP: Refactoring Pseudo CSS Styling
