+++ client/django-formset/SortableSelect.scss
@@ -1,5 +1,6 @@
 django-sortable-select {
 	overflow: hidden scroll;
+	border: 1px solid rgb(118, 118, 118);
 
 	> option {
 		line-height: normal;
