+++ client/django-formset/_dualselector.scss
@@ -13,7 +13,7 @@
 		.control-panel {
 			padding: 1rem 0.5rem;
 		}
-		select {
+		select, django-sortable-select {
 			height: 100%;
 			width: 10rem;
 		}
